+++ internal/repository/index/associated_data.go
