+++ internal/repository/index/associated_data_test.go
