+++ internal/repository/repair_index.go
@@ -3,8 +3,8 @@ package repository
 import (
 	"context"
 
-	"github.com/restic/restic/internal/index"
 	"github.com/restic/restic/internal/pack"
+	"github.com/restic/restic/internal/repository/index"
 	"github.com/restic/restic/internal/restic"
 	"github.com/restic/restic/internal/ui/progress"
 )
